fix: keep bench preset parsing modular
